+++ src/org/itmo/java/lesson3/HW3/HouseNew.java
@@ -5,8 +5,8 @@ public class HouseNew {
         House house1 = new House();
         House house2 = new House();
 
-        house1.setAll(20, 1988, "Хрущевка, 5 этажей");
-        house2.setAll(70, 2025, "небоскреб в проекте");
+        house1.setAll((short) 20, 1988, "Хрущевка, 5 этажей");
+        house2.setAll((short) 70, 2025, "небоскреб в проекте");
 
         System.out.println(house1.toString());
         System.out.println(house2.toString());
